move UiTextColorMarkerField.ts to formfield directory
